repl: move completion logic to internal module
PR-URL: https://github.com/nodejs/node/pull/59889
