+++ lib/internal/repl/utils.js
@@ -14,6 +14,7 @@ const {
   StringPrototypeLastIndexOf,
   StringPrototypeReplaceAll,
   StringPrototypeSlice,
+  StringPrototypeStartsWith,
   StringPrototypeToLowerCase,
   StringPrototypeTrim,
   Symbol,
@@ -49,6 +50,10 @@ const {
   getStringWidth,
   inspect,
 } = require('internal/util/inspect');
+
+const CJSModule = require('internal/modules/cjs/loader').Module;
+
+const vm = require('vm');
 
 let debug = require('internal/util/debuglog').debuglog('repl', (fn) => {
   debug = fn;
@@ -783,6 +788,55 @@ function isObjectLiteral(code) {
     RegExpPrototypeExec(endsWithSemicolonRegExp, code) === null;
 }
 
+const kContextId = Symbol('contextId');
+
+const path = require('path');
+
+function fixReplRequire(replModule) {
+  try {
+    // Hack for require.resolve("./relative") to work properly.
+    replModule.filename = path.resolve('repl');
+  } catch {
+    // path.resolve('repl') fails when the current working directory has been
+    // deleted.  Fall back to the directory name of the (absolute) executable
+    // path.  It's not really correct but what are the alternatives?
+    const dirname = path.dirname(process.execPath);
+    replModule.filename = path.resolve(dirname, 'repl');
+  }
+
+  // Hack for repl require to work properly with node_modules folders
+  replModule.paths = CJSModule._nodeModulePaths(replModule.filename);
+}
+
+let nextREPLResourceNumber = 1;
+// This prevents v8 code cache from getting confused and using a different
+// cache from a resource of the same name
+function getREPLResourceName() {
+  return `REPL${nextREPLResourceNumber++}`;
+}
+
+const globalBuiltins =
+  new SafeSet(vm.runInNewContext('Object.getOwnPropertyNames(globalThis)'));
+
+let _builtinLibs = ArrayPrototypeFilter(
+  CJSModule.builtinModules,
+  (e) => e[0] !== '_' && !StringPrototypeStartsWith(e, 'node:'),
+);
+
+// Note: the `getReplBuiltinLibs` and `setReplBuiltinLibs` are functions used to provide getters and
+//       setters for the `builtinModules` and `_builtinLibs` properties of the repl module and for making
+//       sure that all internal repl modules share the same value, which can potentially be updated by users.
+//       Also note that both `repl.builtinModules` and `repl._builtinLibs` are deprecated, once such properties
+//       are removed these two functions should also be removed as no longer necessary.
+
+function getReplBuiltinLibs() {
+  return _builtinLibs;
+}
+
+function setReplBuiltinLibs(value) {
+  _builtinLibs = value;
+}
+
 module.exports = {
   REPL_MODE_SLOPPY: Symbol('repl-sloppy'),
   REPL_MODE_STRICT,
@@ -792,4 +846,10 @@ module.exports = {
   setupReverseSearch,
   isObjectLiteral,
   isValidSyntax,
+  kContextId,
+  getREPLResourceName,
+  globalBuiltins,
+  getReplBuiltinLibs,
+  setReplBuiltinLibs,
+  fixReplRequire,
 };
